committing code that selects the longest word from the quote
